+++ data-raw/maintrepair_spec.R
@@ -1,4 +1,4 @@
-# FlexFile Data Spec for building database
+# Maintenance and Repair Parts Report Data Spec for building database
 
 library(magrittr)